# openapi/swaggerSpec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "API documentation for My API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/cart/{cartid}/items/{id}": {
      "delete": {
        "summary": "Delete item",
        "description": "Delete an item from a user's cart",
        "parameters": [
          {
            "in": "path",
            "name": "cartid",
            "description": "ID of the cart",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "description": "ID of the product",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the deletion operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "success",
                        "fail",
                        "error"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "description": "The message of the response"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "200": {
        "application/json": {
          "schema": null,
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "enum": [
                "success",
                "fail",
                "error"
              ]
            },
            "data": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "the message of the response"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}